Use draw_idle() everywhere instead of draw()
draw_idle() will perform a draw() when control returns to the Qt
event loop. It has the potential for major performance increases
because multiple calls to draw_idle() functions in a slot will
only result in one draw() when control returns to the event loop.

It resulted in major performance boosts for the brightness and
contrast editor, and I'm hoping it will give performance boosts
elsewhere as well. Just need to double check that we didn't call
draw_idle() on some object that doesn't have that function...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>