Add option to apply pixel solid angle correction
This option currently lives under the "Edit" menu, and it applies to
all three views.

My 24 logical CPUs all run near 100% while it is being computed for
the first time, and it can freeze up for a little while, but after
they have been computed, if they are memoized in hexrd (via
hexrd/hexrd#260), it is much faster afterward, since they do
not need to be computed again.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>