Add the messages widget to progress dialogs
The progress dialogs now contain a messages widget, which will only
be shown if a message is written.

The exec_() method of the progress dialog clears text and resets
the size of the dialog so that the same progress dialog can be used
for multiple runs.

This also fixes an issue where, if an exception occurred during
instantiation of the MainWindow(), there would be no messages printed
since the MessagesWidget was overriding stdout. Now, the traceback
gets printed for sure.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>