Display values in tree view when loaded with hdf5
Before, if the instrument config was loaded with an hdf5 file,
there would be no values displayed in the tree view, since Qt did
not know how to display numpy types.

This converts the data from a numpy type to a native type so that
Qt will know how to display it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>