Distinguish between working directory and directory path
Update variables so that there is a clearer distinction between which represents
the full path to the selected directory and which represents the working
directory that we need to save to the indexing config. Also make sure that we are
saving the parent/stem to the config rather than the full file path.

Signed-off-by: Brianna Major <brianna.major@kitware.com>