Add safety checks for dicts in reflections table
For hexrd's PlaneData class, sometimes the exclusions list can be
longer than the hkl list that gets returned. I specifically noticed
this happening if tThMax gets reduced.

This was resulting in indexing errors in the reflections table, which
put the reflections table in a very unusable state.

To prevent this kind of thing from happening both now and in the future
(if other changes are made), add some safety checks for the dicts. This
will at least prevent the reflections table from becoming unusable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>