Warn the user for too many seed search quaternions
If the seed search yields over 5e7 test orientations, bring up a message
box that shows the user how many test orientations were generated,
along with a warning that it may use up too much memory. The user can
choose whether to proceed or not.

If the user chooses to not proceed, the GUI returns back to the
eta omega maps viewer, so the user can select new options and try again.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>