Capture hexrd logging output as well
In addition to other kinds of stdout and stderr, capture hexrd logging
output as well. In the messages widget, this will print out INFO and
DEBUG levels in green, and WARNING, ERROR, and CRITICAL levels in red.
The default logging level is INFO. The logging output also gets printed
to the console.

Unlike hexrd CLI, this only prints the message, and not the time of
the message. This is because the messages widget is not very big, and
the output looks a little nicer without the times. However, this can be
changed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>