Get packaging to always use the latest hexrd
For release builds, this should use the latest hexrd release.

For pre-release builds, this should use the latest hexrd pre-release.

This constraint is necessary to get the conda packaging to always use
the latest hexrd. Otherwise, conda may choose older hexrd versions,
especially if older hexrd versions allow it to use a newer version of
a different package.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>