Abstract accepting progress to prevent seg fault
It appears that if you are in a thread other than the main thread,
and you try to create a Qt connection with a function on a Qt object
in the main thread, a segmentation fault can occur.

Abstract this away so that the connection to the Qt object is made
only once: when the runner is instantiated. Then, calling "accept_progress()"
emits a signal that causes the progress to be accepted.

This appears to fix the seg fault.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>