Only generate clustering parameters if needed
This adds a property to check if clustering parameters are needed.

If the clustering parameters are not needed, it sets min_samples to 1.

This property will also be useful for re-running the clustering parameters,
since this will show us whether it should be an option in the dialog
or not...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>