Create new HexrdConfig property for unaggregated images
There are a lot of cases throughout the code that need to use the unaggregated
imageseries if it exists or the imageseries dict if it does not. The unagg_images
property has been updated to return the imageseries dict if the unaggregated
images do not exist. This also adds an agg_images property that returns a boolean
indicating whether or not we are currently viewing an aggregated imageseries.

Signed-off-by: Brianna Major <brianna.major@kitware.com>