Add a thermal factor editor to allow tensor U
This allows the thermal factor (U/B) to be either a tensor or a scalar.

As a scalar, the user can modify U in its spinbox as normal. If the
user wishes to convert it to a tensor, they must double-click the
spinbox, and then the thermal factor editor will appear, allowing
them to convert it into a tensor.

If U is a tensor, the spinbox text will just display "Tensor", and
the user must click on it in order to bring up the thermal factor
editor, in order to edit it.

This was tested via round-tripping (creating a tensor U, writing
it out to a materials file, resetting the application, and then
loading the materials file again, to ensure that the tensor U
was saved and loaded properly).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>