Move "Subtract Minimum" to an intensity correction
This removes the "Subtract Minimum" option from the transform dialog,
and moves it into the "Edit"->"Intensity Corrections" menu.

This means that the internal re-workings of "Subtract Minimum" have
changed as well. It used to be an operation in a `ProcessedImageSeries`,
but now it is only applied to the image the user is currently viewing,
just like the other intensity corrections do.

We may want to consider changing the intensity corrections to also use
operations in `ProcessedImageSeries` so the intensity corrections will
be applied to the whole image series if it gets passed around. Currently,
if the whole image series gets passed somewhere (like the HEDM workflow),
none of the intensity corrections will be applied.

These re-workings have also fixed #979, since "Subtract Minimum" is now
using the intensity correction framework, rather than its own framework.

Fixes: #979

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>