Improve performance of applying polar masks
Previously, the image processing steps of rescaling the intensities
and applying the snip algorithm would be performed when polar masks
were applied (this included toggling their visibility).

Now, the snipped image is cached so that when polar mask settings
change, only the polar masks need to be re-applied, and the snip
algorithm does not need to run again. This significantly improves
performance, as the snip algorithm can take a long time for high
resolution images.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>