Increase log min to 1.e-7
For some datasets (such as pxrdip), if we do a min of 1.e-8,
matplotlib will round this to zero and then raise an exception later
stating that the min cannot be zero.

However, if we use 1.e-7, matplotlib does not round this to zero, and
an exception does not get raised.

I don't know why this is the case, but maybe we can keep it at 1.e-7
for now.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>