Use checkbox to convert between scalars/tensors
This is a little more intuitive for the user to use, when converting
all of the thermal factors from scalars to tensors, or back.

Toggling the checkbox to checked will convert all thermal factors
to tensors. Toggling the checkbox to unchecked will convert all
thermal factors to scalars. The user is, however, still allowed to
change individual thermal factors to tensors and back.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>