Add polar binary erosion to remove edge artifacts
This adds an "Apply erosion?" checkbox to the polar tab of the
image mode widget, which can only be checked if "Apply SNIP?" is
checked.

If checked, a binary erosion is applied to the panel buffer mask, and
it uses 2x the snip width for the structure. This sets edge artifacts
to zero.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>