Add ellipsoidal ranges for laue overlays
Ellipsoidal ranges are closer to reality than rectangular ranges, so
the laue overlays now use ellipsoidal ranges by default, but the user
can still change them to use rectangular ranges via the laue overlay
editor.

The ellipses are generated via the equation of an ellipse in polar
space coordinates. `panel.angles_to_cart()` and
`panel.panel.cartToPixel()` are used to convert the ellipses for
display in the cartesian and raw views.

It appears that masking via the ellipse ranges is working properly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>