Add X-Ray Energy Selection via a Table
This adds a button to the form view which allows the user to pick an
energy (or multiple energies to average) from a table.

The `characteristic_xray_energies.h5` from hexrd is being used to
generate this table.

The table is currently sorted so that atomic symbols come first,
and are sorted in alphabetical order, followed by atomic numbers.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>