Re-create SNIP preview plot each time
Re-using the figure and the axis seemed to be a reasonable thing to
do. However, it is causing issues for when the user re-shows it.
On Linux, the user can't change the color map the second time they
show it. On Mac, it doesn't re-appear at all.

Instead of re-using, we will now re-create the figure from scratch each
time. It is not very resource intensive to do this, and figuring out
why we can't re-use the figure and axis do not seem worth the effort.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>