Use conda-forge instead of anaconda channels
This is an attempt to use the `conda-forge` channel instead of the
`defaults` and `anaconda` channels, for all packages.

This will hopefully fix the issue where we had to downgrade scikit-learn
because it would fail to install on Github Actions on Windows. It makes
more sense for us to install these packages from conda-forge since the
respective project developers actually maintain the packages on
conda-forge (whereas their packages on the anaconda/defaults channels
are maintained by anaconda).

We will first need to make sure the packaging and tests pass, and then
move on to test the packages manually.

If we run into crashing issues on Mac with openblas, as we have in the
past, we may be able to resolve it by setting the environment variable
`OPENBLAS_NUM_THREADS=1`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>