Block enter key press accept/reject for dialogs
Previously, enter key presses would be blocked if the QDialog
received them. However, sometimes, the QPushButtons in the
QDialogButtonBox in the QDialog would receive the key press event
instead, thus bypassing our event filter. This would result in
the dialog being accepted/rejected accidentally, upon accidental
enter key press, and thus the issue we tried to prevent would not
always be fixed.

Now, the key event filter is installed on both the QDialogs and
on QPushButtons that follow this parent/child scheme:

QDialog -> QDialogButtonBox -> QPushButton

The event filter must be installed both on the QDialog and the
QPushButtons because it is currently unpredictable (to me,
at least) which one will receive the key press event, and if the
QDialog receives the key press event, it DOES NOT forward the
event to the QPushButton, but rather "clicks" the QPushButton.

This appears to fix all cases that I found.

Fixes: #1004

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>