Pass mask name through signal
Originally it was assumed that the currentRow for the table would be accurate
each time a cell widget was toggled or clicked. This behavior is not guaranteed
though and on some platforms seems to introduce bugs. To make sure we always
have the correct value the key is now passed through the widget signal.

Signed-off-by: Brianna Major <brianna.major@kitware.com>