Move sorting logic into GrainsTableView
It makes more sense for the sorting logic to lie there, especially
because it prevents us from having to duplicate the logic for the
grains table view that will be in the fit grains options dialog.

This also starts display a grains table view in the fit grains options
dialog, which can visualize spots, but needs more testing, and also
needs options for the user to select tolerances to use.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>