Add a general unique list editor
The unique list editor allows users to edit a general list of strings
by doing things such as renaming, re-ordering, deleting, and copying.
Signals get fired for each kind of modification.

This editor will be used for editing the list of materials in the
materials panel.

This list editor does not allow duplicates.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>