Add materials list editor
This adds a `MaterialsListEditor` dialog that is used to edit the list
of materials.

The user can change the ordering of the materials, copy, delete, rename,
and import materials.

This "takes over" the functionality of the tool button in the materials
panel. That button no longer displays menu items (which included adding a
material, importing a material, and deleting a material). But now, when
the user clicks on the tool button, the materials list editor appears.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>