Convert to native type in TreeItem.set_data()
Rather than having to keep tracking where np.ndarray and np.generic
values are in all of the various classes, just go ahead and perform
the conversion in one place: TreeItem.set_data(). This makes adding
new classes easier.

The conversion is necessary for Qt to be able to display the value.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>