Allow NaNs to be set in ScientificDoubleSpinBox
This adds some special behavior that allows NaNs to be set.
Conversion to/from NaNs seems to work properly.

This also includes a fix to ValueColumnDelegate.setEditorData()
to ensure that `setValue()` is called. It appears that we have to
do this in order for Qt to call our override of `setValue()`. I'm
guessing this is because internally Qt is calling `setProperty("value", ...)`
instead of `setValue()`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>