Add picks tree view with export
This adds a button to the line picker dialog that allows the user
to see a tree view of the picks they have chosen, and export these
picks.

The user is currently not able to edit the points, but that is a
feature that is planned for the near future.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>