Add point editing and drawing
Points can now be edited, and selected points will be drawn in the canvas.

This now hides overlays and drawn points before showing the picks tree
view, and the picks tree view takes over overlay and point drawing.

Also, the calibration runner updates its overlay picks with the modifications
that the user makes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>