Always add line when moving to next powder overlay
The fact that we were not adding a new matplotlib line when moving
to the next powder overlay (after not inserting points in one) would
cause some issues with re-drawing the lines.

Now, we always add a line, even when no data was entered. We also
have removed the "ring_data" variable as it is unused and obsolete.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>