Add ability to export/import WPPF parameters
This adds buttons to the WPPF options dialog so that the user can
export/import the WPPF parameters as an HDF5 file.

The importing performs some validation to make sure all of the
keys in the imported file match the keys that are currently in the
GUI. The keys will not match if the user has different materials
selected, and/or a different peak shape method.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>