Separate out the actual loading of masks from h5py group
Importing masks now has two functions: import_masks is called if the import
button is used in the mask manager dialog and will read a selected file and
create an h5py group. The load_masks function takes in an h5py group, parses
the data and creates the masks.

Signed-off-by: Brianna Major <brianna.major@kitware.com>