Add support for saving configuration in HDF5
We convert the configuration to YAML and then use a custom
loader/dumper to allow us to include numpy types in the
HDF5 file in separate datasets. The high level YAML configuration
is saved as a string dataset. We use the SafeLoader for loading
so we have whitelisted a few safe python types.