Add auto point picking as calibration option
When performing a calibration, if a powder overlay is about to be
used, a window now appears asking the user if they wish to auto pick
points, or hand pick points.

If the user chooses to hand pick points, it proceeds with the line
picker dialog as it did before.

If the user chooses to auto pick the points, a dialog appears that
gets the user to select options for the auto picking. When the user
clicks "Ok", auto picking is performed, and if successful, a dialog
will appear showing the user the picks, and allowing the user to
modify the picks.

The user can add, delete, and re-pick by hand points that were picked.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>