Make mask loading more robust
For some reason, when we load the masks from a state file instead of
just importing the masks, the `_visible` items come back as bytes instead
of strings, and the mask data is not loaded.

This fixes the issue so that the mask visibilities are converted to
strings properly, and the `_visible` items come out as strings.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>