Add a bunch of state file loading/saving pieces
This adds saving/loading the material, instrument, configuration, and
masks to the state file.

A lot of pieces seem to be working correctly, but I'm sure there will
be some bugs or things that we missed, as these are some big changes.

This currently skips the data, just due to developing time constraints.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>