Step according to singleStep in scientific spinbox
Previously, we would step by the digit to the left of the decimal
place (whether we were in scientific notation, or standard notation).
However, this was not that useful, because most numbers were not large
enough or small enough to be in scientific notation, and we had no
control over the step size.

Instead, let's step according to the step size, so that we are able to
better control what it will be. Now, we can set the `singleStep` property
to control the step size.

Fixes: #1065

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>