Add ability to load crystal parameters from grains
In the calibration crystal editor (found in both the Laue overlay
editor and the Rotation Series overlay editor), there is now a
"Load Crystal Parameters" button at the bottom, that allows the user
to load crystal parameters from grains. It gives the user the choice
to display grains from the fit grains output, find orientations output,
or from a file. The user may select a grain in the table and click
"OK", and the crystal parameters will be populated with the parameters
from the selected grain.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>