Add multi-grain HEDM calibration
For multi-grain HEDM calibration, the user must first create rotation series
overlays until they have the same number as the number of grains they want
to use in the calibration.

When they start the HEDM calibration, they will select that number of grains
from the grains table. Next, in the HEDM Calibration Options dialog, they will
be able to pair their selected grains with the rotation series overlays. They
are also able to view the grains.

Finally, the calibration is ran and results for individual grains can be
viewed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>