Refactor overlays to be classes instead of dicts
The overlay dicts have become more and more complicated as time has
gone on, and new synchronizations between their values and other
parts of the program (such as the image series) have made them even
more complicated to maintain.

As such, it has become a good time to convert them to be classes
instead of dicts, where the classes can internally keep up with any
synchronizations that are needed.

This will also hopefully fix a bug we were encountering where an
overlay was being modified to have a refinements list from another
overlay. This shouldn't be possible with objects.

There is a new format for serializing the overlays, but backward
compatible readers are also now used so that the transition is
seamless.

Since the overlays are used in many, many parts of the code, we
really need to do some thorough testing to find and fix any bugs
this refactoring may cause.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>