Catch IndexError on mouse move
This error would happen occasionally in the cartesian and polar views
when the mouse is moved close to an edge of the canvas. It appears that
it was happening because the x_data for the event was slightly over the
max, which led to an out-of-bounds index being computed. This would
then result in an IndexError.

This error is relatively harmless, but it may worry users, so catch it
to prevent it from appearing.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>