Use image view data for B&C editor histogram
This makes the B&C editor histogram's "Auto" button _significantly_
more useful for the Cartesian and Polar views, since it actually now
works on them properly.

This commit added a image_view_loaded signal that contains the new
image view data so that the main window can connect to it and update
the B&C editor, so that the histogram can stay up-to-date.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>