Add scaling options to color map editor
This adds a few scaling options to the color map editor:

1. sqrt
2. log
3. log-log-sqrt

These are selected via a combo box. If one is selected, the data
will be transformed using the appropriate scaling before the image
data is passed to matplotlib. The transformed image is then rescaled to
have the same data range as the original image.

NOTE: the data is transformed for viewing purposes only. Any kinds of
computations with the data still runs only on unscaled data.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>