Fix B&C editor histogram plot
It was previously creating a plot of a histogram of the histogram,
rather than just plotting the histogram itself. Fix that by using
matplotlib's `bar()` function.

This also fixes an issue where there were x margins that made the
line and the bars off with respect to each other. Those x margins
are now removed, and the line and the bars now appear to match up.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>