Add support for '\r' in the MessagesWidget
In the console, '\r' causes the cursor to go to the beginning of the
line. This is what tools like `tqdm` use in order to update a status
bar.

Add support for '\r' by going back to the beginning of the line in the
text edit when '\r' is encountered.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>