Add optional labels to Laue overlays
This adds a "Labels" section to the Laue overlay editor where the user
may choose to label the Laue overlays by either hkls or energy. The user
is able to control the label offsets in both x and y directions (which
also accept negative values).

The user is also able to modify the label style via the overlay style
editor dialog.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>