Fix HexrdConfig().has_images
This should only return `True` when there are images loaded that are
not dummy images. It was previously returning `True` when there were
dummy images loaded.

This also fixes the brightness and contrast editor to be disabled when
there are dummy images (because otherwise, it will produce a not
very clear error if you try to use it).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>