Add indexing results dialog
The indexing results dialog plots the indexing results on top of the
eta omega maps. It is important for determining whether the indexing
was successful.

Similar to the eta omega maps viewer, the user is able to select an hkl
to view. However, in this dialog, the user may also select grains to
plot on top of the eta omega maps as well.

Pressing "OK" will continue on to the fit grains options dialog. Pressing
"Cancel" will stop the workflow.

Fixes: #952

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>