Force material name to match upon adding
This was causing a bug when copying the materials in the materials list
editor, because the copy of the material was given a new name, but the
name on the material object itself did not match.

This means that places where the name on the material object itself was
used (such as the reflections table) would be used incorrectly.

Fix this by forcing newly added materials to have the name that matches
their key in the dict.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>