Add c/a ratio editor
The c/a ratio editor is only enabled for space groups 75 - 194, where
the c/a ratio is important.

If the c/a ratio is modified, the `a` parameter will be modified to
reflect the ratio.

The ratio may also be fixed by checking `Fix?`. If checked, then if
`a` or `c` is modified, the other parameter will also be modified
automatically in order to keep the `c/a` ratio constant.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>