Only load the "CeO2" material by default
This modifies hexrdgui to only load the "CeO2" material by default,
which improves start-up time because some computations that are ran
when the materials are created take up some time.

The other default materials are available in the material list editor
via a "Import from Defaults" button. If the user clicks this button,
they may select other default materials to import.

If the program exits cleanly, the names of the default materials that
were loaded get saved, so that when hexrdgui is restarted, these default
materials will also be loaded again. Additionally, if any overlays were
saved that are using default material names, these default materials will
also be loaded on start-up.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>