Add hand-picked indexing method
This allows the user to hand pick the grains by first selecting the
hand-picked quaternion method in the eta omega maps viewer, then
right-clicking on a spot. This will create a list of fibers from 0
to 360 degrees that intersect with the spot. A slider can then be
moved to determine the angle at which the fiber intersects with other
spots as well. This fiber can then be added to the list of quaternions.

If the user utilizes this method, the workflow then proceeds directly
to fit grains.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>