Add fiber step option to hand picking
This adds a fiber step option to the hand picking indexing, which is
always synchronized with the fiber step option in the seed search
quaternion method.

If the user modifies the fiber step, the fibers get regenerated with
the new step. The current angle will also be rounded to the nearest
valid angle.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>