Save last hkl index for generating fibers
This fixes a bug where, if the user generated fibers based upon a
clicked position, and then they switched to a different hkl index,
and then they modified the fiber step, the newly generated fibers
would be incorrect.

This is because when the user modifies the fiber step, it regenerates
the fibers using the new step. The hand picked coordinates used to
generate the fibers was previously saved and would be used to generate
the fibers with the new steps. However, the hkl index used to generate
the fibers was not saved, and thus the fibers would be generated
with an incorrect hkl index.

Save and use the last hkl index to avoid this issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>