Use reduced hkl ids for generating angles
This fixes a bug where, when we skipped some hkls in the hkl selection,
the angles that were generated would not be correct.

This was because the output list of this function would be of the full
length of the hkls, rather than just the length of the reduced hkls.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>