Save/load refinement statuses in the state file
These were previously not being saved/loaded because we are saving the
instrument config differently from how we do so with QSettings. In state
files, we are using hexrd's HEDMInstrument to write out the instrument
config, but this does not include refinement statuses.

This PR adds a property to HexrdConfig() to create/set a dict of
refinement statuses, and this dict is used to save/load the
refinement statuses from the state file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>