Calculate index for selected point
This fixes the issue we were running into earlier, where the indices
would not always match correctly.

It does so by transforming all of the 3D data into 2D points (as
displayed on the screen), and then finding the data point that is
the closest to the location of the mouse click.

It seems to work for different orientations of the graph, and for
different sortings of the grains table as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>