Migrate PowderCalibrator to the one in hexrd
We were previously using a custom PowderCalibrator class for the
"Laue and Powder" calibration in hexrdgui. However, this class is
not maintained and no longer works properly.

Migrate over to using the PowderCalibrator from hexrd. This one
is more maintained and updated.

The migration required modifying the data input, since the
PowderCalibrator from hexrd expects the input data to be in a different
format.

There may be some options we should be using and are not using yet.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>