Make conversion to cartesian consistent
Because the instrument gets updated for every iteration, the conversion
to cartesian coordinates was being changed for every iteration. This
would cause the least squares to diverge significantly.

Fix it so that we use a consistent conversion to cartesian coordinates
by using a copy of the original HEDMInstrument.

This also makes a few small modifications so that it is more consistent
with hexrd's InstrumentCalibrator.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>