Add option to display the beam marker
This adds an option in the main window to display the beam marker, that
defaults to `True`.

If `True`, the beam marker will be drawn in all three views.

The user may edit the style of the beam marker by toggling it on. This
makes the style editor for the beam marker appear.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>