Add ability to save lineout as xy or csv
If the image mode is "Polar", and the user selects
"File" -> "Export" -> "Current View", and they
name their file with a ".xy" or ".csv" extension, the azimuthal
integral plot is written in XY or CSV format, respectively. This
follows the previous pattern where the user could write out to NPZ
format by specifying a ".npz" extension.

However, there may certainly be some room to make this more visible
or user friendly. I'm open to suggestions. Maybe right-clicking the
azimuthal integral plot in the polar view should bring up a context
menu with the option to export it, and then they choose which file
format (XY or CSV)? They could likewise do this for exporting other
views.

This PR is also a good first step.

Fixes: #1188
Fixes: #1232

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>