Add a viewer for the panel buffer
This adds a button to the "Configure Panel Buffer" widget that allows
the user to view numpy panel buffers via a pop-up imshow.

If there is no numpy file selected, it shows the current panel buffer.
If there is a numpy file selected, it shows the panel buffer for the
numpy file.

The button is disabled if the panel buffer is in border mode rather than
in numpy mode.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>