Add image calculator
Similar to the image calculator in ImageJ, this allows you to perform
arithmetic or logical operations between images.

In hexrdgui, the user first selects a detector to use. Then they select
the operation to perform. Finally, they select the operand (technically,
the second operand) via selecting a file.

When "OK" is clicked, some validation is performed, with some informative
error messages in case any errors occur. Additionally, if the operand is
an image series with more than one frame, a dialog will appear asking the
user to select an aggregation to perform.

Once the operation is complete, the image in the selected detector will
be replaced with the result of the operation.

Note: this feature is only enabled in the GUI if the loaded image series
is either aggregated or only has a length of 1.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>