Remove use of QSignalBlocker in most code
It appears that we can't always rely on python's reference counting
system to delete `QSignalBlocker` (or lists of `QSignalBlocker`) at
the end of method calls. As such, we should convert over to using the
`block_signals()` context manager from `hexrd.ui.utils`.

This commit moves all code over to that.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>