Add ability to delete grains in some dialogs
This adds the ability to delete grains in the fit grains options
dialog (before fit grains is ran), and in the fit grains results
dialog (after fit grains is ran).

The user may delete grains by right-clicking them in the table and
clicking "Delete grain(s)". Note that it is currently set up to not
allow the user to delete all grains.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>