Fix Fast Powder calibration in Cartesian view
If you ran the fast powder calibration within the cartesian view,
you would encounter a `KeyError` saying that the detector key is missing
from the dict when `overlay.hkls[det_key]` was called.

This is because the overlay in the cartesian view uses a fake instrument
that does not have the right detector keys.

However, we still need the correct hkls to save the calibration picks
correctly (since the calibration picks assume an order that is consistent
with the correct hkls).

Thus, now, if an overlay was generated using the Cartesian view, a temporary
overlay will be created in the raw view that has the correct hkls, and these
hkls will be used for saving the calibration data.

Fixes: #1261

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>