Clear unaggregated images also during images clear
Previously, we were not clearing unaggregated images in
`HexrdConfig().clear_images()`. Clear them as well so that all of the
images data is fully cleared.

This issue resulted in a bug where, if you had an aggregated image series
loaded, and then you loaded a new instrument config, the unaggregated images
would be used and would have extra incorrect detectors in it.

The unaggregated images are only used if an aggregated image series is loaded,
in order to make a backup of the unaggregated images.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>