Add Cartesian input option for beam vector
In the Form View, there is now a combo box that allows the user to
select the input type for the beam vector. It defaults to polar, but
can be changed to Cartesian.

If the user modifies the beam vector in Cartesian mode, a note is
displayed that shows what the vector gets normalized to.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>