Add editor for beam vector magnitude
This is hidden by default in the tree view, but can be enabled via
a checkbox in the form view.

Editing the beam vector magnitude modifies the beam vector values,
and editing beam vector values modifies the beam vector magnitude.

If the beam vector is infinite, then the beam vector is normalized.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>