Add tth distortion for polar images
If there are powder overlays present with tth distortion active, and
we are in the polar view, then some options will be present for applying
tth distortion to the polar view.

There is a checkbox where the user can select whether to apply tth
distortion to the polar view. If checked, the user must then select an
overlay (that has tth distortion) to use.

The tth distortion of the selected overlay will be applied to the polar
view, and any other overlays with tth distortion will be offset
accordingly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>