Do not switch to raw upon apply pinhole mask
The signal `raw_masks_changed` was previously causing the raw image mode to
be loaded. This wasn't a problem before, because that signal was only
emitted while the user was in the raw view.

However, the pinhole mask is now always being saved in the raw view, and if
the user applies it while in the polar view, the view gets switched
automatically to the raw view.

Now, the `raw_masks_changed` signal simply causes whatever view is current
to update, so that we don't automatically switch to the raw view.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>